contract verification api moved to api project
